requirements.txt
pexpect==4.8.0
pyserial==3.5
PyYAML==5.4.1
pytest==6.2.3
RPi.GPIO==0.7.0
